Add unit test in spatial mesh
